Switch to game scrapes to using the team code - and speaking of which change acronym to code (it's more of an abbreviation I suppose but that feels clunky to write). This gets around the issue of having teams like Arizona/Phoenix be referred to in different ways.
Make things that are not all available at initial game scrape
nullable, like Arena, attendance, and game id.

Add a "shell" for easy inspection of things (really I'm just
dropping to a pdb prompt - obviously longer term we need to fix
this but this works for now).

This adds all* games for a season to the database in populate.

* Well, not precisely all, but close enough for now.